ROW_COUNT() functionality and tests

+++ sql/plan/row_update_accumulator.go
@@ -229,7 +229,7 @@ func (a *accumulatorIter) Close(ctx *sql.Context) error {
 	}
 
 	result := a.updateRowHandler.okResult()
-	ctx.SetLastQueryInfo(sql.RowCount, result.RowsAffected)
+	ctx.SetLastQueryInfo(sql.RowCount, int64(result.RowsAffected))
 	return nil
 }
